Navigation Drawer, revert bendable pitch additions, remove redundent layouts
Additional activity_main layouts are redundant at the time of this commit. Landscape is disabled for phones and on tablets landscape is the same as portrait

What was previously in the options menu has been moved to a Navigation bar. I also switched to a theme that has the action bar hidden as it is unnecessary to have.

The previous bendable pitch additions were causing a lot of problems with sound production. They have been removed until the SoundHandler class is refactored to support it.