Implement option to disable pitch bending, fix hint reset bug
When the user disables pitch bending the slider correct slider positions are randomized. Playback is given the offset to produce the correct note.

The hints didn't reset when dismissing the dialog or selecting no, the final hints would then pop up after the first check. It has been fixed so that the streak is always reset to 0. New slider positions are now chosen when a chord is correct.